+++ account_product_fiscal_classification/__init__.py
@@ -1,4 +1,4 @@
-# -*- encoding: utf-8 -*-
+# -*- coding: utf-8 -*-
 from . import model
 
 import logging
