+++ account_product_fiscal_classification/model/__init__.py
@@ -1,5 +1,7 @@
-# -*- encoding: utf-8 -*-
+# -*- coding: utf-8 -*-
 from . import product_template
+from . import account_product_fiscal_classification_model
+from . import account_product_fiscal_classification_template
 from . import account_product_fiscal_classification
 from . import wizard_change_fiscal_classification
 from . import wizard_account_product_fiscal_classification
