+++ account_product_fiscal_classification/model/account_product_fiscal_classification.py
@@ -1,4 +1,4 @@
-# -*- encoding: utf-8 -*-
+# -*- coding: utf-8 -*-
 ##############################################################################
 #
 #    Account Product - Fiscal Classification module for Odoo
@@ -24,51 +24,6 @@
 
 from openerp import models, fields, api, _
 from openerp.exceptions import ValidationError
-
-
-class AccountProductFiscalClassificationModel(models.AbstractModel):
-    """Fiscal Classification model of customer and supplier taxes.
-    This classification is used to create Fiscal Classification
-    and Fiscal Classification template."""
-    _name = 'account.product.fiscal.classification.model'
-    _MAX_LENGTH_NAME = 256
-
-    # Field Section
-    code = fields.Char()
-
-    name = fields.Char(
-        size=_MAX_LENGTH_NAME, required=True, select=True, translate=True)
-
-    description = fields.Text()
-
-    active = fields.Boolean(
-        default=True,
-        help="If unchecked, it will allow you to hide the Fiscal"
-        " Classification without removing it.")
-
-
-class AccountProductFiscalClassificationTemplate(models.Model):
-    """Fiscal Classification model of customer and supplier taxes.
-    This classification is used to create Fiscal Classification
-    and Fiscal Classification template."""
-    _name = 'account.product.fiscal.classification.template'
-    _inherit = 'account.product.fiscal.classification.model'
-
-    purchase_tax_ids = fields.Many2many(
-        comodel_name='account.tax.template',
-        relation='fiscal_classification_template_purchase_tax_rel',
-        column1='fiscal_classification_id', column2='tax_id',
-        string='Purchase Taxes', oldname="purchase_base_tax_ids", domain="""[
-            ('parent_id', '=', False),
-            ('type_tax_use', 'in', ['purchase', 'all'])]""")
-
-    sale_tax_ids = fields.Many2many(
-        comodel_name='account.tax.template',
-        relation='fiscal_classification_template_sale_tax_rel',
-        column1='fiscal_classification_id', column2='tax_id',
-        string='Sale Taxes', oldname="sale_base_tax_ids", domain="""[
-            ('parent_id', '=', False),
-            ('type_tax_use', 'in', ['sale', 'all'])]""")
 
 
 class AccountProductFiscalClassification(models.Model):
