+++ account_product_fiscal_classification/model/account_product_fiscal_classification_model.py
@@ -2,8 +2,8 @@
 ##############################################################################
 #
 #    Account Product - Fiscal Classification module for Odoo
-#    Copyright (C) 2014-Today GRAP (http://www.grap.coop)
-#    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
+#    Copyright (C) 2015 -Today Akretion (http://www.akretion.com)
+#    @author Renato Lima (https://twitter.com/renatonlima)
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
@@ -20,4 +20,27 @@
 #
 ##############################################################################
 
-from . import tests
+
+from openerp import models, fields, api, _
+from openerp.exceptions import ValidationError
+
+
+class AccountProductFiscalClassificationModel(models.AbstractModel):
+    """Fiscal Classification model of customer and supplier taxes.
+    This classification is used to create Fiscal Classification
+    and Fiscal Classification template."""
+    _name = 'account.product.fiscal.classification.model'
+    _MAX_LENGTH_NAME = 256
+
+    # Field Section
+    code = fields.Char()
+
+    name = fields.Char(
+        size=_MAX_LENGTH_NAME, required=True, select=True, translate=True)
+
+    description = fields.Text()
+
+    active = fields.Boolean(
+        default=True,
+        help="If unchecked, it will allow you to hide the Fiscal"
+        " Classification without removing it.")
